Fix bad import into core
