+++ packages/lexical-utils/src/index.ts
@@ -9,7 +9,6 @@
 
 import {$cloneWithProperties} from '@lexical/selection';
 import {
-  $copyNode,
   $createParagraphNode,
   $getRoot,
   $getSelection,
@@ -19,6 +18,7 @@ import {
   $isRootOrShadowRoot,
   $isTextNode,
   $setSelection,
+  $splitNode,
   DEPRECATED_$isGridSelection,
   EditorState,
   ElementNode,
@@ -27,6 +27,8 @@ import {
   LexicalNode,
 } from 'lexical';
 import invariant from 'shared/invariant';
+
+export {$splitNode};
 
 export type DFSNode = Readonly<{
   depth: number;
@@ -386,51 +388,4 @@ export function $wrapNodeInElement(
   node.replace(elementNode);
   elementNode.append(node);
   return elementNode;
-}
-
-export function $splitNode(
-  node: ElementNode,
-  offset: number,
-): [ElementNode | null, ElementNode] {
-  let startNode = node.getChildAtIndex(offset);
-  if (startNode == null) {
-    startNode = node;
-  }
-
-  invariant(
-    !$isRootOrShadowRoot(node),
-    'Can not call $splitNode() on root element',
-  );
-
-  const recurse = (
-    currentNode: LexicalNode,
-  ): [ElementNode, ElementNode, LexicalNode] => {
-    const parent = currentNode.getParentOrThrow();
-    const isParentRoot = $isRootOrShadowRoot(parent);
-    // The node we start split from (leaf) is moved, but its recursive
-    // parents are copied to create separate tree
-    const nodeToMove =
-      currentNode === startNode && !isParentRoot
-        ? currentNode
-        : $copyNode(currentNode);
-
-    if (isParentRoot) {
-      currentNode.insertAfter(nodeToMove);
-      return [
-        currentNode as ElementNode,
-        nodeToMove as ElementNode,
-        nodeToMove,
-      ];
-    } else {
-      const [leftTree, rightTree, newParent] = recurse(parent);
-      const nextSiblings = currentNode.getNextSiblings();
-
-      newParent.append(nodeToMove, ...nextSiblings);
-      return [leftTree, rightTree, nodeToMove];
-    }
-  };
-
-  const [leftTree, rightTree] = recurse(startNode);
-
-  return [leftTree, rightTree];
 }
