+++ packages/lexical/src/index.ts
@@ -141,6 +141,7 @@ export {
   $nodesOfType,
   $setCompositionKey,
   $setSelection,
+  $splitNode,
   isSelectionWithinEditor,
 } from './LexicalUtils';
 export {$isDecoratorNode, DecoratorNode} from './nodes/LexicalDecoratorNode';
